Post and Comment status and flagging
Introduced basic functionality to support different statuses for posts
and comments, to properly filter them when listing. Also, added a
flagging system. For every flag assigned to a comment or post, the
author loses karma. If the post or comment is flagged more than 5
times, its status is set to flagged. Requires migration.