Initial support for themes
It is now possible to create a themes folder inside of the application
folder, where custom themes can be stored. Switching to a theme that is
not dillo currently requires the admin templates folder to be copied or
symlinked into the theme folder. This is not ideal and will be solved
in the future when a fully integrated admin backend is implemented.