Tweak to thumbnail() and original_image() in Post model
If imgur_client is defined, now we store the full link to the original
image (instead of the db). We generate thumbnails just by parsing the
link and adding the size suffix. This greatly reduces the Imgur API
consumption. Introduce automatic migration of old image ids to full
links.

The main downside is that if Imgur changes their url schema, we have to
deal with it on the db side.