Introducing Community Theme Color
Foundations for using a custom color for certain accents.

Idea is to have communities assign this color to their liking.
Only front-end so far, it's missing the back-end work to set such color.

Uses CSS variables, which are supported in every major browser nowadays.
That excludes IE (Edge is ok).

Fallback is $color-primary (defined in config sass files)