Fix HTML not being rendered on posts anymore.
Emoji parsing was using text instead of html, because we were serving pure html
before we couldn't tell. Now that HTML is done via the "markdowned" filter
it, emoji needs html. This emoji parsing needs cleaning as it is a bit all over
the place now, added it to my to-do.

Fixes #101