Make comments generic
Before this commit, comments were only related to Posts. Since we
moved to an Entity-based model, it makes sense to relate Comments to
any model.
This commit drops the post field and introduces a GenericForeignKey
in its place, with a migration to swap Post ids with their Generic
representation.

In the process of adding this feature, a number of unit tests were
updated/fixed.